memory: Add persistent_allocation
persistent_allocation is an allocation which can hold on to more memory than
what is required for the current number of elements stored.

This mimics std::vector behavior and is intended to reduce reallocation
overhead. It exposes a reserve()/shrink_to_fit()/clear() interface that is
similar (but not identical) to std::vector, in addition to the plain allocation
interface.