timer: Allow host timers to be copied
Firepony heavily relies on copying host timers around. Since the host
timer is still effectively POD, reinstate the copy constructor for host
timers.

Note that GPU timers remain very much uncopiable. This is a significant
change in semantics between the two, so this change is hopefully
temporary.

Work on #63 (which contains more details on this problem)