pointer: Fix definition of thrust_const_iterator_type
thrust_const_iterator_type needs to be defined as an iterator to a const
in order for the wrapper to work properly. This was preventing code that
calls t_begin()/t_end() on const Lift pointers from compiling.

Also adds a unit test to ensure this doesn't break again.

Fixes #77